requirements.txt
alembic==1.4.2
appnope==0.1.0
Babel==2.8.0
backcall==0.1.0
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
decorator==4.4.2
dominate==2.5.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
guess-language-spirit==0.5.3
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
traitlets==4.3.3
urllib3==1.25.9
visitor==0.1.3
wcwidth==0.1.9
Werkzeug==1.0.1
WTForms==2.2.1
